feat: Add installation scripts and utility functions for project setup
- Created a new script `composer.sh` to automate the creation of essential script files in the `lib` directory.
- Introduced `config.sh` for project configuration, including versioning and licensing information.
- Added `info.sh` for displaying banners and installation summaries.
- Developed `install.sh` to handle the installation process, including build and download options.
- Implemented `install_funcs.sh` for installation-related functions, including UPX installation and PATH management.
- Created `platform.sh` to detect the operating system and architecture for binary releases.
- Added `utils.sh` for utility functions, including logging and temporary directory management.
- Introduced `validate.sh` for validating Go version and checking dependencies.
- Updated `CLI_VERSION` to version 1.3.0 and removed the obsolete `current.go` file.
- Enhanced `semantic.go` to support dynamic module name and alias configuration via environment variables.
