+++ .github/pull_request_template.md
@@ -1,0 +1,31 @@
+# Pull Request
+
+## Description
+
+<!-- Please include a summary of the change and which issue is fixed. Also include relevant motivation and context. List any dependencies that are required for this change. -->
+
+Fixes #
+
+## Checklist
+
+Please check all items that apply before requesting a review:
+
+- [ ] My code follows the project’s coding style and guidelines
+- [ ] I have tested my changes locally
+- [ ] I have updated the documentation as needed
+- [ ] All new and existing tests passed
+- [ ] I have added necessary tests where applicable
+
+## Related Issue(s)
+
+<!-- List all issues this PR closes or is related to. Example: Closes #123, Related to #456 -->
+
+## Screenshots (if applicable)
+
+<!-- If your change affects the UI, please include screenshots here. Otherwise, you can remove this section. -->
+
+---
+
+**Before approving this PR, reviewers must verify:**
+- [ ] All checklist items are completed
+- [ ] Linked related issues are valid and referenced above
