+++ NOTICE.md
@@ -1,0 +1,15 @@
+# NOTICE
+
+This software is licensed under the MIT License. Below are additional notes on usage and attribution:
+
+## Attribution Requirement (Optional)
+
+- When distributing or using this software, please provide credit to the original author(s) in one or more of the following ways:
+- Retain the copyright notice: `Copyright (c) 2025 Rafael Mori`.
+- Include a link to the original project repository or website.
+
+## Acknowledgment
+
+This project was developed with the goal of enhancing usability and providing open access to its features.
+
+For further information about the license and terms of use, please refer to the `LICENSE` file included with this project.
