+++ version/CLI_VERSION
@@ -1,1 +1,1 @@
-v1.0.8
+v1.3.0
